fix: remove thread pool in favor of one thread per request
Disable support for keep alive by removing the threadpool and leveraring
a single thread per request. This is required cause keep-alive can
exhaust the threadpool especially if it is quiet small. Using a
threadpool or other async I/O implementations could be regained after
support for it in the standard library gets finished.